perf: optimize PgmData and reduce memory usage / 优化PgmData并减少内存占用
- Apply linear scan for small ranges in PgmData::get (jdb_pgm/src/data.rs)
- Simplify predict_in_seg with branchless clamp (jdb_pgm/src/pgm.rs)
- Change Segment indices from usize to u32 (jdb_pgm/src/types.rs)
- Reduce LUT memory usage by 50% via u32 (jdb_pgm/src/build.rs)

- 在PgmData::get中对小范围应用线性扫描（jdb_pgm/src/data.rs）
- 用无分支clamp简化predict_in_seg（jdb_pgm/src/pgm.rs）
- 将Segment索引类型从usize改为u32（jdb_pgm/src/types.rs）
- 通过u32将LUT内存减少50%（jdb_pgm/src/build.rs)

+++ jdb_pgm/src/types.rs
@@ -40,8 +40,8 @@ pub struct Segment<K: Key> {
   pub max_key: K,
   pub slope: f64,
   pub intercept: f64,
-  pub start_idx: usize,
-  pub end_idx: usize,
+  pub start_idx: u32,
+  pub end_idx: u32,
 }
 
 /// Index statistics
